feat: List DBs actually on the target connection
When using the dump, restore, or migrate subcommands without specifying
a database for one of the connections. The user will be prompted with a
list of databases queried from the connection. This was previously just
using a hardcoded list for testing but now actually queries the
databases on the connection.